% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getResult}
\alias{getResult}
\title{Title}
\usage{
getResult(
  sample,
  data_type,
  group_tag,
  response_tag,
  customer,
  sig_file = NULL,
  exp_file = NULL,
  save_abun = "Sample_abundance",
  save_group = "Group_abundance",
  save_icb = "ICB_response",
  save_plot = "Group_comparison.png",
  save_bar = "Abundance_barplot.png"
)
}
\arguments{
\item{save_bar}{}
}
\value{

}
\description{
Title
}
