% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuCellAI.R
\name{ImmuCellAI}
\alias{ImmuCellAI}
\title{Title}
\usage{
ImmuCellAI(
  sample,
  data_type,
  group_tag,
  response_tag,
  customer,
  sig_file = NULL,
  exp_file = NULL
)
}
\arguments{
\item{exp_file}{}
}
\value{

}
\description{
Title
}
\examples{

}
