% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuCellAI.R
\name{sample_ratio}
\alias{sample_ratio}
\title{Title}
\usage{
sample_ratio(data, marker_exp, marker_tag_mat, data_type)
}
\arguments{
\item{data_type}{}
}
\value{
sample expression deviation matrix
}
\description{
Title
}
