% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuCellAI.R
\name{self_sig}
\alias{self_sig}
\title{self-build reference file}
\usage{
self_sig(sig_file)
}
\arguments{
\item{sig_file}{self-build cell type gene signature file.}
}
\value{
A list of cell type gene signatures.
}
\description{
self-build reference file
}
\details{
Please input the expression matrix separated by tab, rownames of the file must be gene symbol.
}
